interesting...am I missing something.  or what am I missing?
I have stopped collecting data, but maybe only now is the work beginning.  I have attached some Excel files with graphs.  Take a look.

Gravity and mass.  Are they constant?  This is a Boolean yes/no question.  “Well, they are sort of constant” If they are “sort of constant” they that means they are not constant!
  
Reach out to me!   mrphysh@juno.com